Naive conversion with common functions extracted
