+++ vulnfeeds/cmd/nvd-cve-osv/main_test.go
@@ -4,6 +4,7 @@ import (
 	"reflect"
 	"testing"
 
+	"github.com/google/osv/vulnfeeds/cmd/common"
 	"github.com/google/osv/vulnfeeds/cves"
 	"github.com/google/osv/vulnfeeds/internal/testutils"
 )
@@ -11,8 +12,8 @@ import (
 func TestReposFromReferences(t *testing.T) {
 	type args struct {
 		CVE         string
-		cache       VendorProductToRepoMap
-		vp          *VendorProduct
+		cache       common.VendorProductToRepoMap
+		vp          *common.VendorProduct
 		refs        []cves.Reference
 		tagDenyList []string
 	}
@@ -26,14 +27,14 @@ func TestReposFromReferences(t *testing.T) {
 			args: args{
 				CVE:   "CVE-2023-0327",
 				cache: nil,
-				vp:    &VendorProduct{"theradsystem_project", "theradsystem"},
+				vp:    &common.VendorProduct{"theradsystem_project", "theradsystem"},
 				refs: []cves.Reference{
 					{
 						Source: "cna@vuldb.com",
 						Tags:   []string{"Patch", "Third Party Advisory"},
 						Url:    "https://github.com/saemorris/TheRadSystem/commit/bfba26bd34af31648a11af35a0bb66f1948752a6"},
 				},
-				tagDenyList: RefTagDenyList,
+				tagDenyList: common.RefTagDenyList,
 			},
 			wantRepos: []string{"https://github.com/saemorris/TheRadSystem"},
 		},
@@ -42,14 +43,14 @@ func TestReposFromReferences(t *testing.T) {
 			args: args{
 				CVE:   "CVE-2025-0211",
 				cache: nil,
-				vp:    &VendorProduct{"campcodes", "school_faculty_scheduling_system"},
+				vp:    &common.VendorProduct{"campcodes", "school_faculty_scheduling_system"},
 				refs: []cves.Reference{
 					{
 						Source: "cna@vuldb.com",
 						Tags:   []string{"Exploit", "Third Party Advisory"},
 						Url:    "https://github.com/shaturo1337/POCs/blob/main/LFI%20in%20School%20Faculty%20Scheduling%20System.md"},
 				},
-				tagDenyList: RefTagDenyList,
+				tagDenyList: common.RefTagDenyList,
 			},
 			wantRepos: []string(nil),
 		},
@@ -66,7 +67,7 @@ func TestReposFromReferences(t *testing.T) {
 						Url:    "https://git.musl-libc.org/cgit/musl/commit/?id=c47ad25ea3b484e10326f933e927c0bc8cded3da",
 					},
 				},
-				tagDenyList: RefTagDenyList,
+				tagDenyList: common.RefTagDenyList,
 			},
 			wantRepos: []string{"https://git.musl-libc.org/git/musl"},
 		},
@@ -83,7 +84,7 @@ func TestReposFromReferences(t *testing.T) {
 						Url:    "https://github.com/dwyl/hapi-auth-jwt2/issues/111",
 					},
 				},
-				tagDenyList: RefTagDenyList,
+				tagDenyList: common.RefTagDenyList,
 			},
 			wantRepos: []string{"https://github.com/dwyl/hapi-auth-jwt2"},
 		},
@@ -91,7 +92,7 @@ func TestReposFromReferences(t *testing.T) {
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
 			testutils.SetupGitVCR(t)
-			if gotRepos := ReposFromReferences(tt.args.CVE, tt.args.cache, tt.args.vp, tt.args.refs, tt.args.tagDenyList); !reflect.DeepEqual(gotRepos, tt.wantRepos) {
+			if gotRepos := common.ReposFromReferences(tt.args.CVE, tt.args.cache, tt.args.vp, tt.args.refs, tt.args.tagDenyList, Logger); !reflect.DeepEqual(gotRepos, tt.wantRepos) {
 				t.Errorf("ReposFromReferences() = %#v, want %#v", gotRepos, tt.wantRepos)
 			}
 		})
@@ -100,20 +101,20 @@ func TestReposFromReferences(t *testing.T) {
 
 func Test_maybeUpdateVPRepoCache(t *testing.T) {
 	type args struct {
-		cache VendorProductToRepoMap
-		vp    *VendorProduct
+		cache common.VendorProductToRepoMap
+		vp    *common.VendorProduct
 		repos []string
 	}
 	tests := []struct {
 		name      string
 		args      args
-		wantCache VendorProductToRepoMap
+		wantCache common.VendorProductToRepoMap
 	}{
 		{
 			name: "Test with no cache",
 			args: args{
 				cache: nil,
-				vp:    &VendorProduct{"avendor", "aproduct"},
+				vp:    &common.VendorProduct{"avendor", "aproduct"},
 				repos: []string{"https://github.com/google/osv.dev"},
 			},
 			wantCache: nil,
@@ -121,45 +122,45 @@ func Test_maybeUpdateVPRepoCache(t *testing.T) {
 		{
 			name: "Test with an empty cache",
 			args: args{
-				cache: VendorProductToRepoMap{},
-				vp:    &VendorProduct{"avendor", "aproduct"},
+				cache: common.VendorProductToRepoMap{},
+				vp:    &common.VendorProduct{"avendor", "aproduct"},
 				repos: []string{"https://github.com/google/osv.dev"},
 			},
-			wantCache: VendorProductToRepoMap{
-				VendorProduct{"avendor", "aproduct"}: []string{"https://github.com/google/osv.dev"},
+			wantCache: common.VendorProductToRepoMap{
+				common.VendorProduct{"avendor", "aproduct"}: []string{"https://github.com/google/osv.dev"},
 			},
 		},
 		{
 			name: "Test with an empty cache and an unusable repo",
 			args: args{
-				cache: VendorProductToRepoMap{},
-				vp:    &VendorProduct{"avendor", "aproduct"},
+				cache: common.VendorProductToRepoMap{},
+				vp:    &common.VendorProduct{"avendor", "aproduct"},
 				repos: []string{"https://github.com/vendor/repo"},
 			},
-			wantCache: VendorProductToRepoMap{},
+			wantCache: common.VendorProductToRepoMap{},
 		},
 		{
 			name: "Test with an existing cache",
 			args: args{
-				cache: VendorProductToRepoMap{
-					VendorProduct{"avendor", "aproduct"}: []string{"https://github.com/google/osv.dev"},
+				cache: common.VendorProductToRepoMap{
+					common.VendorProduct{"avendor", "aproduct"}: []string{"https://github.com/google/osv.dev"},
 				},
-				vp:    &VendorProduct{"avendor", "aproduct"},
+				vp:    &common.VendorProduct{"avendor", "aproduct"},
 				repos: []string{"https://github.com/google/osv-scanner"},
 			},
-			wantCache: VendorProductToRepoMap{
-				VendorProduct{"avendor", "aproduct"}: []string{"https://github.com/google/osv.dev", "https://github.com/google/osv-scanner"},
+			wantCache: common.VendorProductToRepoMap{
+				common.VendorProduct{"avendor", "aproduct"}: []string{"https://github.com/google/osv.dev", "https://github.com/google/osv-scanner"},
 			},
 		},
 		{
 			name: "Test with an empty cache adding two values",
 			args: args{
-				cache: VendorProductToRepoMap{},
-				vp:    &VendorProduct{"avendor", "aproduct"},
+				cache: common.VendorProductToRepoMap{},
+				vp:    &common.VendorProduct{"avendor", "aproduct"},
 				repos: []string{"https://github.com/google/osv.dev", "https://github.com/google/osv-scanner"},
 			},
-			wantCache: VendorProductToRepoMap{
-				VendorProduct{"avendor", "aproduct"}: []string{"https://github.com/google/osv.dev", "https://github.com/google/osv-scanner"},
+			wantCache: common.VendorProductToRepoMap{
+				common.VendorProduct{"avendor", "aproduct"}: []string{"https://github.com/google/osv.dev", "https://github.com/google/osv-scanner"},
 			},
 		},
 	}
@@ -167,7 +168,7 @@ func Test_maybeUpdateVPRepoCache(t *testing.T) {
 		t.Run(tt.name, func(t *testing.T) {
 			testutils.SetupGitVCR(t)
 			for _, repo := range tt.args.repos {
-				maybeUpdateVPRepoCache(tt.args.cache, tt.args.vp, repo)
+				common.MaybeUpdateVPRepoCache(tt.args.cache, tt.args.vp, repo)
 			}
 			if !reflect.DeepEqual(tt.args.cache, tt.wantCache) {
 				t.Errorf("maybeUpdateVPRepoCache() have %#v, wanted %#v", tt.args.cache, tt.wantCache)
